+++ src/ContextException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Facile\Sentry\Log;
-
-class ContextException extends \RuntimeException
-{
-}